Updates the plan's do_install and vendoring python deps
The project deps are now vendered in the root of the package.
This was to flatten the path and make it more sane. The
resulting path makes it easy to run the vendored pip instead
of sourcing the activate file.

The resulting `flask` binary is not on the path for the run
hook so I need to specify the full path. I chose to do that
for the application file instance and the flask binary.

Signed-off-by: Franklin Webber <franklin@chef.io>